realizando conexão com o banco de dados e parando de usar dados mockados

+++ src/controllers/termo.controller.js
@@ -1,32 +1,81 @@
 const termosService = require('../services/termo.service');
+const mongoose = require('mongoose');
 
-const findAllTermosController = (req, res) => {
-  const termos = termosService.findAllTermosService();
+const findAllTermosController = async (req, res) => {
+  const termos = await termosService.findAllTermosService();
+
+  if (termos.length == 0) {
+    return res.status(404).send({
+      message: 'Não existe nenhum termo cadastrado vamos começar? ;)',
+    });
+  }
   res.send(termos);
 };
 
-const findByIdTermosController = (req, res) => {
-  const parametroId = Number(req.params.id);
-  const escolhaTermo = termosService.findByIdTermoService(parametroId);
+const findByIdTermosController = async (req, res) => {
+  const parametroId = req.params.id;
+
+  if (!mongoose.Types.ObjectId.isValid(parametroId)) {
+    return res.status(400).send({ message: 'Id inválido, tente novamente...' });
+  }
+
+  const escolhaTermo = await termosService.findByIdTermoService(parametroId);
+
+  if (!escolhaTermo) {
+    return res
+      .status(404)
+      .send({ message: 'Termo não encontrado no Guide Dev ;(' });
+  }
+
   res.send(escolhaTermo);
 };
 
-const createTermoController = (req, res) => {
+const createTermoController = async (req, res) => {
   const termo = req.body;
-  const newTermo = termosService.createTermoService(termo);
-  res.send(newTermo);
+
+  if (!termo || !termo.descricao || !termo.logo || !termo.ano_de_criacao) {
+    return res
+      .status(400)
+      .send({ message: 'Envie todos os campos do novo termo, por favor! >:(' });
+  }
+
+  const newTermo = await termosService.createTermoService(termo);
+  res.status(201).send(newTermo);
 };
 
-const editTermoController = (req, res) => {
-  const idParam = Number(req.params.id);
+const editTermoController = async (req, res) => {
+  const idParam = req.params.id;
+
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    return res.status(400).send({ message: 'Id inválido, tente novamente...' });
+  }
+
   const termoEdit = req.body;
-  const editedTermo = termosService.editTermoService(idParam, termoEdit);
+
+  if (
+    !termoEdit ||
+    !termoEdit.descricao ||
+    !termoEdit.logo ||
+    !termoEdit.ano_de_criacao
+  ) {
+    return res
+      .status(400)
+      .send({ message: 'Envie todos os campos do termo, por favor! >:(' });
+  }
+
+  const editedTermo = await termosService.editTermoService(idParam, termoEdit);
   res.send(editedTermo);
 };
 
-const deleteTermoController = (req, res) => {
+const deleteTermoController = async (req, res) => {
   const idParam = req.params.id;
-  termosService.deleteTermoService(idParam);
+
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    return res.status(400).send({ message: 'Id inválido, tente novamente...' });
+  }
+
+  await termosService.deleteTermoService(idParam);
+
   res.send({ message: 'Termo deletado com sucesso do dicionário! :)' });
 };
